+++ libc/stdlib.h
@@ -6,13 +6,14 @@ FrostOS implementation of stdlib.h
 #define STDLIB_H
 
 #include <stdint.h>
+#include <stddef.h>
 #include "./mem.h"
 #include "./string.h"
 
 char *itoa(int64_t value, char * str, int base);
 double pow (double x, int y);
-void llimit(char *src, uint32_t n, char a);
-void rlimit(char *src, uint32_t n, char a);
+void llimit(char *src, size_t n, char a);
+void rlimit(char *src, size_t n, char a);
 char *ftoa(double value, char *str, int base, int precision);
 
 #endif
